test(workbar): cover the workbar's tabs in one Storybook group
The workbar reads four bridges at once (tasks, artifacts, inspector,
sessions), which is why the shell had no story: the panel this PR
rebuilds was reviewable only by launching the app. Product/Session
Workbar mocks all four in one decorator and renders each tab through the
real shell — seven states across tasks, files and trace, including the
three the e2e fixture cannot reach because its session has never run a
turn.

Rendering through the shell is what lets the group replace rather than
join the per-panel stories. FIDELITY.md asks for exactly that: import the
wrapper instead of retyping it, and let a story earn its place by
rendering pixels no other story renders. So Product/Artifact Pane and the
hand-written boxes that would have propped up separate trace and
task-ledger groups are gone, along with the Artifact Pane frame's
`data-maka-e2e-fixture` — it pauses every animation and transition via
base.css, something only the E2E harness does, and
capability-audit-strip.stories.tsx already removed it for that reason.

product-smoke-manifest.json names none of the retired story ids, so the
manifest pass is unaffected. The smoke run here is 68 manifest checks and
79 catalog renders.
